Use CV ports on internal modules.
git-svn-id: http://svn.drobilla.net/lad/trunk/ingen@4301 a436a847-0d15-0410-975c-d299462d15a1

+++ src/server/internals/Note.hpp
@@ -82,7 +82,7 @@ private:
 		SampleCount time;
 	};
 
-	float note_to_freq(int num);
+	float note_to_freq(uint8_t num);
 	void free_voice(ProcessContext& context, uint32_t voice, FrameTime time);
 
 	Raul::Array<Voice>* _voices;
